Move error functions to its own util
