+++ shell/utils/error.js
@@ -104,3 +104,29 @@ export function exceptionToErrorsArray(err) {
     return [err];
   }
 }
+
+
+/**
+ * Imported from path-to-regexp
+ * @param {*} err 
+ * @returns 
+ */
+export const normalizeError = (err) => {
+  let message;
+
+  if (!(err.message || typeof err === 'string')) {
+    try {
+      message = JSON.stringify(err, null, 2);
+    } catch (e) {
+      message = `[${ err.constructor.name }]`;
+    }
+  } else {
+    message = err.message || err;
+  }
+
+  return {
+    ...err,
+    message,
+    statusCode: (err.statusCode || err.status || (err.response && err.response.status) || 500)
+  };
+};
